+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.1
+
+* Add a `Future.cast` extension method.
+
 ## 0.4.0
 
 * Removed `RestartableTimer`. `package:async` already provides such a class.